refactor: files' structur and styling for Form component
