+++ task-components/src/pages/MainPage.tsx
@@ -1,38 +1,31 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import Header from '../components/Header';
 import Search from 'components/Search';
-import { ApiCardContent } from 'utils/types';
 import ApiCardsList from 'components/ApiCardsList';
 import Loader from 'components/Loader';
 import { useAppDispatch, useAppSelector } from 'app/hooks';
-import { addCards, selectCards } from './mainPageSlice';
-import { useGetCharactersByNameQuery } from './ApiSlice';
-import { selectSearch } from 'components/searchSlice';
+import { addCards, selectCards } from '../slices/mainPageSlice';
+import { useGetCharactersByNameQuery } from '../slices/ApiSlice';
+import { selectSearch } from 'slices/searchSlice';
 
 const MainPage: React.FC = () => {
   const savedCards = useAppSelector(selectCards);
   const dispatch = useAppDispatch();
+  const saved = useAppSelector(selectSearch);
   // const [characters, setCharacters] = useState<ApiCardContent[]>([]);
-  // const [isLoading, setLoading] = useState(true);
   const savedValue = useAppSelector(selectSearch);
   const query = savedValue || '';
-  const { data, error, isLoading } = useGetCharactersByNameQuery(query, {
+  const { data, error, isFetching } = useGetCharactersByNameQuery(query, {
     refetchOnMountOrArgChange: true,
   });
-
-  // useEffect(() => {
-  //   return () => {
-  //     dispatch(addCards(characters));
-  //   };
-  // });
 
   return (
     <div>
       <Header name="Home"></Header>
       <Search></Search>
       {data && <ApiCardsList cards={data.results}></ApiCardsList>}
-      {isLoading && <Loader />}
-      {!data && <div>Something went wrong...</div>}
+      {isFetching && <Loader />}
+      {(!data || error) && <div className="home-error">Oops... Something went wrong...</div>}
     </div>
   );
 };
