+++ task-components/src/styles/navigation.css
@@ -11,5 +11,5 @@
 	text-decoration: none;
 	font-size: 1.4rem;
 	font-weight: 500;
-	transition: 0.3s;
+	transition: 0.4s;
 }
